pyproject.toml
[project]
name = "study-roulette"
version = "0.1.0"
description = "A Stable Randomizer for Study Recruitment Links"
readme = "README.md"
authors = [
    { name = "Nate Vack", email = "njvack@wisc.edu" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "uvicorn[standard]>=0.40.0",
]

[build-system]
requires = ["uv_build>=0.9.14,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
    "ty>=0.0.13",
]
